Remove warning about unclosed connections.
When xml2::read_html() failed to read an HTML file, it didn't close
the connection. This led to an unnecessary warning that came
randomly later. The code it was assigned to changed. By passing a
connection directly to read_html(), I am able to close it afterwards
if the function errors.

Here is the logic that decides how to process the input:
https://github.com/r-lib/xml2/blob/685bfff4a6dd85a069ce73db6fb560a2b065a6e5/R/xml_parse.R#L106

read_html() was already using suppressWarnings(), which is why my
attempts to suppress the warnings also failed.
https://github.com/r-lib/xml2/blob/685bfff4a6dd85a069ce73db6fb560a2b065a6e5/R/xml_parse.R#L77