Fix file protocol for testing local files on Windows.
For xml2::url_parse() to correctly parse the file protocol, there
must be an additional slash in from the Windows drive. This is
consistent with the link in a browser when opening a local HTML
file.

However, for xml2::read_html(), the file path must be a standard
Windows path, so the slash preceding the drive must be removed prior
to reading.