Optimize formatting for GitHub-Flavored Markdown (GFM).
For ordered sublists, I have to switch to "1)", which gets rendered
as "i." in GFM. This looks dumb in pandoc Markdown, because the
numbering just starts over in the sublist.

https://github.github.com/gfm/#list-items

For the README, I can't use a CSS code chunk. That only has an effect
when rendered directly to HTML with pandoc. Thus I use <img> tags to
explicitly control the height of the example favicons.