Stop scheduled runs of `R CMD check`
This was mainly to regularly test the ability to grab favicons
from important websites. But this was too noisy, and only ever
exposed temporary problems with accessing the website, and never
about my code's behavior. Now that I stopped running the at_home
tests, there is no reason to regularly run the checks if I have
made no changes.